+++ R/convert_tocsv.R
@@ -5,6 +5,8 @@
 #'
 #' @inheritParams lr_get_spec
 #'
+#' @inherit lr_get_spec details
+#'
 #' @return Convert input files to csv and invisibly return the list of created
 #' file paths
 #'
@@ -16,12 +18,18 @@
 #'
 #' @importFrom tools file_path_sans_ext
 #' @importFrom utils write.csv
+#' @importFrom future.apply future_lapply
+#' @importFrom progressr with_progress progressor
 #'
 #' @export
 lr_convert_tocsv <- function(where = NULL, ext = "txt", decimal = ".",
-                             sep = NULL, subdir = FALSE,
-                             cores = getOption("mc.cores", 2L),
+                             sep = NULL, subdir = FALSE, cores = NULL,
                              ignore.case = TRUE, overwrite = FALSE) {
+
+  if (!missing(cores)) {
+    warning("'cores' argument is deprecated. See ?future::plan for more info ",
+            "about how you can choose your parallelisation strategy.")
+  }
 
   if (is.null(where)) {
     warning("Please provide a valid location to read and write the files.",
@@ -45,18 +53,17 @@ lr_convert_tocsv <- function(where = NULL, ext = "txt", decimal = ".",
 
   files <- file.path(where, file_names)
 
-  if (cores > 1 && .Platform$OS.type == "windows") {
-    cores <- 1L
-    message('Parallel processing not available in Windows; "cores" set to 1.\n')
-  }
-
   message(nb_files, " files found")
 
-  tmp <- pbmclapply(files, function(x)
-    tryCatch(spec2csv_single(x, decimal = decimal, sep = sep,
-                             overwrite = overwrite),
-             error = function(e) NULL),
-    mc.cores = cores)
+  with_progress({
+    p <- progressor(along = files)
+    tmp <- future_lapply(files, function(x) {
+      p()
+      tryCatch(spec2csv_single(x, decimal = decimal, sep = sep,
+                               overwrite = overwrite),
+               error = function(e) NULL)
+    })
+  })
 
   whichfailed <- which(vapply(tmp, is.null, logical(1)))