+++ R/get_spec.R
@@ -7,36 +7,55 @@
 #'
 #' @param where Folder in which files are located (defaults to current working
 #' directory).
-#' @param ext File extension to be searched for, without the "."
-#' (defaults to `txt`).
+#' @param ext File extension to be searched for, without the "." (defaults to
+#' `txt`). You can also use a character vector to specify multiple file
+#' extensions.
 #' @param lim A vector with two numbers determining the wavelength limits to be
 #' considered (defaults to `c(300, 700)`).
 #' @param subdir Should subdirectories within the `where` folder be included in
 #' the search? (defaults to `FALSE`).
 #' @param subdir.names Should subdirectory path be included in the name of the
 #' spectra? (defaults to `FALSE`).
-#' @param cores Number of cores to be used. If greater than 1, import will use
-#' parallel processing (not available in Windows).
+#' @param cores deprecated. See [future::plan()] for more details on how to
+#' customise your parallelisation strategy.
 #' @param ignore.case Should the extension search be case insensitive? (defaults
 #' to `TRUE`)
+#' @param interpolate Boolean indicated whether spectral data should be
+#' interpolated and pruned at every nanometer. Note that this option can only
+#' work if all input data samples the same wavelengths. Defaults to `TRUE`.
+#'
+#' @details
+#' You can customise the type of parallel processing used by this function with
+#' the [future::plan()] function. This works on all operating systems, as well
+#' as high performance computing (HPC) environment. Similarly, you can customise
+#' the way progress is shown with the [progressr::handlers()] functions
+#' (progress bar, acoustic feedback, nothing, etc.)
 #'
 #' @return A data.frame, containing the wavelengths in the first column and
 #' individual imported spectral files in the subsequent columns.
-#' Reflectance values are interpolated to the nearest wavelength integer.
 #'
 #' @export
 #'
-#' @importFrom pbmcapply pbmclapply
 #' @importFrom tools file_path_sans_ext
 #' @importFrom stats approx
+#' @importFrom future.apply future_lapply
+#' @importFrom progressr with_progress progressor
+#'
+#' @seealso [pavo::getspec()]
 #'
 #' @examples
 #' lr_get_spec(system.file("testdata", package = "lightr"), ext = "jdx")
 #'
 lr_get_spec <- function(where = getwd(), ext = "txt", lim = c(300, 700),
-                     decimal = ".", sep = NULL, subdir = FALSE,
-                     subdir.names = FALSE, cores = getOption("mc.cores", 2L),
-                     ignore.case = TRUE) {
+                        decimal = ".", sep = NULL, subdir = FALSE,
+                        subdir.names = FALSE, cores = NULL,
+                        ignore.case = TRUE, interpolate = TRUE) {
+
+  if (!missing(cores)) {
+    warning("'cores' argument is deprecated. See ?future::plan for more info ",
+            "about how you can choose your parallelisation strategy.",
+            call. = FALSE)
+  }
 
   extension <- paste0("\\.", ext, "$", collapse = "|")
 
@@ -66,25 +85,33 @@ lr_get_spec <- function(where = getwd(), ext = "txt", lim = c(300, 700),
 
   range <- seq(lim[1], lim[2])
 
-  if (cores > 1 && .Platform$OS.type == "windows") {
-    cores <- 1L
-    message('Parallel processing not available in Windows; "cores" set to 1.\n')
-  }
-
   message(nb_files, " files found; importing spectra:")
 
-  gsp <- function(ff) {
+  if (!interpolate) {
+    gsp <- function(ff) {
+      dispatch_parser(ff, decimal = decimal, sep = sep)[[1]]
+    }
+  } else {
+    gsp <- function(ff) {
+      df <- dispatch_parser(ff, decimal = decimal, sep = sep)[[1]]
 
-    df <- dispatch_parser(ff, decimal = decimal, sep = sep)[[1]]
+      # Prevent approx from filling a complete gap in the range of interest.
+      bounds <- which(df$wl >= lim[1] & df$wl <= lim[2])
+      df <- df[c(min(bounds)-1, bounds, max(bounds)+1), ]
 
-    interp <- approx(df[, "wl"], df[, "processed"],
-                     xout = range, ties = "ordered")$y
+      approx(df[, "wl"], df[, "processed"],
+             xout = range, ties = "ordered")$y
+    }
   }
 
-  tmp <- pbmclapply(files, function(x)
+  with_progress({
+  p <- progressor(along = files)
+  tmp <- future_lapply(files, function(x) {
+    p()
     tryCatch(gsp(x),
-             error = function(e) NULL
-    ), mc.cores = cores)
+             error = function(e) NULL)
+    })
+  })
 
   whichfailed <- which(vapply(tmp, is.null, logical(1)))
 
@@ -102,12 +129,26 @@ lr_get_spec <- function(where = getwd(), ext = "txt", lim = c(300, 700),
     specnames <- specnames[-whichfailed]
   }
 
-  tmp <- do.call(cbind, tmp)
+  if (interpolate) {
+    tmp <- do.call(cbind, tmp)
 
-  final <- cbind(range, tmp)
+    final <- cbind(range, tmp)
+  } else {
+    if (length(unique(lapply(tmp, function(x) x[, "wl"])))!=1) {
+      stop("'interpolate = FALSE' can only work if all input files sample the ",
+           "same wavelengths.", call. = FALSE)
+    }
+    final <- do.call(cbind, lapply(tmp, function(x) x[, "processed"]))
+
+    final <- cbind(tmp[[1]][, "wl"], final)
+  }
 
   colnames(final) <- c("wl", specnames)
   final <- as.data.frame(final)
+
+  final <- final[final$wl <= lim[2] & final$wl >= lim[1],]
+
+  # TODO: should uninterpolated spectra be marked as rspec objects?
   class(final) <- c("rspec", "data.frame")
 
   return(final)