+++ R/parse_jdx.R
@@ -1,6 +1,7 @@
-#' Parse OceanOptics JCAMP-DX (.jdx) file
+#' Parse OceanInsight JCAMP-DX (.jdx) file
 #'
-#' Parse OceanOptics JCAMP-DX (.jdx) file. <https://oceanoptics.com/>
+#' Parse OceanInsight (formerly OceanOptics) JCAMP-DX (.jdx) file.
+#' <https://www.oceaninsight.com/>
 #'
 #' @inheritParams lr_parse_generic
 #'