+++ inst/doc/batch_import.R
@@ -17,6 +17,10 @@ head(res)
 res <- lr_get_spec(where = "data", ext = "procspec", subdir = TRUE, ignore.case = FALSE)
 
 ## -----------------------------------------------------------------------------
+res <- lr_get_spec(where = "data/heliomaster", ext = "jdx", interpolate = FALSE)
+head(res)
+
+## -----------------------------------------------------------------------------
 res <- lr_get_metadata(where = "data", ext = c("trt", "procspec"), subdir = TRUE)
 head(res)