+++ inst/doc/true_example.R
@@ -1,3 +1,6 @@
+params <-
+list(pavo = TRUE)
+
 ## ----setup, include = FALSE---------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
@@ -14,7 +17,7 @@ knitr::include_graphics("puffin-small.jpg")
 raw_files <- lr_get_spec("data/puffin", ext = "ProcSpec")
 txt_files <- lr_get_spec("data/puffin", ext = "txt")
 
-## ---- fig.show='hold'---------------------------------------------------------
+## ---- fig.show='hold', eval=params$pavo---------------------------------------
 library(pavo)
 plot(raw_files, main = "Raw ProSpec files")
 plot(txt_files, main = "Exported txt files")
@@ -22,11 +25,11 @@ plot(txt_files, main = "Exported txt files")
 ## -----------------------------------------------------------------------------
 all.equal(raw_files, txt_files, tol = 1e-4)
 
-## -----------------------------------------------------------------------------
+## ---- eval=params$pavo--------------------------------------------------------
 beak <- procspec(raw_files, opt = "smooth")
 summary(beak, subset = c("B2", "H5"))
 
-## -----------------------------------------------------------------------------
+## ---- eval=params$pavo--------------------------------------------------------
 vis_beak <- vismodel(beak, visual = "avg.uv", achromatic = "ch.dc")
 tcs_beak <- colspace(vis_beak)
 coldist(tcs_beak, achromatic = TRUE)